UI 1/UI/skills/obdv/pyproject.toml
[project]
name = "OBDV Strategy"
version = "0.1.0"
description = ""
authors = [{ name = "John Doe", email = "john.doe@composabl.com" }]
dependencies = [
    "composabl-core",
    "numpy"
]

[composabl]
type = "skill-teacher"
entrypoint = "obdv.teacher:UPSTeacher"
